Moved page components to Pages directory under src
